packages/python/langsmith-client/pyproject.toml
[project]
name = "langsmith-client"
version = "0.1.0"
description = "LangSmith Control Plane API client and CLI utilities"
requires-python = ">=3.12,<3.13"
dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "python-dotenv>=1.1.0",
    "python-decouple>=3.8",
    "pydantic>=2.0.0",
]

[project.scripts]
langsmith-build = "langsmith_client.tools.build:build"
langsmith-deploy-docker = "langsmith_client.tools.deploy_docker:cli"
langsmith-deploy-github = "langsmith_client.tools.deploy_github:cli"
langsmith-workspaces = "langsmith_client.tools.list_workspaces:cli"
langsmith-listeners = "langsmith_client.tools.list_listeners:cli"
langsmith-projects = "langsmith_client.tools.projects:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/langsmith_client"]
